+++ Sources/UI/CreateGroupCell.swift
@@ -1,0 +1,65 @@
+import UIKit
+
+class CreateGroupCell: UICollectionViewCell {
+    private var image: UIImageView = {
+        let view = UIImageView()
+        view.backgroundColor = UIColor.brandColor.withAlphaComponent(0.5)
+        view.tintColor = .brandColor
+        view.clipsToBounds = true
+        view.layer.masksToBounds = true
+        view.contentMode = .scaleAspectFill
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
+    }()
+
+    private var title: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = .rounded(forTextStyle: .caption2, weight: .semibold)
+        label.textColor = .label
+        label.textAlignment = .center
+        return label
+    }()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        backgroundColor = .clear
+
+        contentView.translatesAutoresizingMaskIntoConstraints = false
+
+        title.text = NSLocalizedString("new_group", comment: "")
+
+        contentView.addSubview(title)
+        contentView.addSubview(image)
+
+        NSLayoutConstraint.activate([
+            contentView.leftAnchor.constraint(equalTo: leftAnchor),
+            contentView.rightAnchor.constraint(equalTo: rightAnchor),
+        ])
+
+        NSLayoutConstraint.activate([
+            image.heightAnchor.constraint(equalToConstant: frame.size.width),
+            image.widthAnchor.constraint(equalToConstant: frame.size.width),
+            image.topAnchor.constraint(equalTo: contentView.topAnchor),
+            image.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
+        ])
+
+        NSLayoutConstraint.activate([
+            title.topAnchor.constraint(equalTo: image.bottomAnchor, constant: 4),
+            title.leftAnchor.constraint(equalTo: contentView.leftAnchor),
+            title.rightAnchor.constraint(equalTo: contentView.rightAnchor),
+        ])
+    }
+
+    required init?(coder _: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        image.image = UIImage(systemName: "plus", withConfiguration: UIImage.SymbolConfiguration(weight: .semibold))
+        image.layer.cornerRadius = frame.width / 2
+    }
+}